// src/enum/TipoMenu.enum.ts
export enum TipoMenu{
    DASHBOARD = 'dashboard', //São os painéis de visualização. Serão filtrados para ficar em posição de destaque
    MENU = 'menu', //São pontos que não redirecionarão para outra página, apenas abrirão para mostrar os itens
    ITEM = 'item' //Redirecionarão o usuário para as páginas
}